+++ HISTORY.rst
@@ -11,6 +11,7 @@ History
   not specified, the GT element will contain a dot.  Prior to this release, the behavior was to emit the 
   ALT allele index.  The old behavior can be retained by setting ``--vcfFailedSnpGt 1``
 * Changed the setup to require PyVCF version 0.6.7 or higher.  It will automatically upgrade if necessary.
+* Added error checking after running SamTools and VarScan to detect missing, empty, or erroneous output files.
 
 
 0.5.0 (2016-01-19)